+++ subcommands/use.js
@@ -5,17 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const path = require('path')
 const chalk = require('chalk')
 const { prompt } = require('inquirer')
-const { configstore } = require('../configstore')
+const { headLessConfigStore } = require('../configstore')
 const { feedback } = require('../utils/cli-feedback')
 const { listSpaces } = require('../utils/spacesUtils')
 const { BB_CONFIG_NAME } = require('../utils/constants')
 const ConfigFactory = require('../utils/configManagers/configFactory')
+const BlockConfigManager = require('../utils/configManagers/blockConfigManager')
+
+
 
 /**
- * Prompt question set the selected answers to config space details
+ * Prompt question set the seleted answers to config space details
  * @param {import('../utils/jsDoc/types').spaceDetails} Data
  */
 const promptAndSetSpace = async (Data) => {
@@ -31,32 +33,31 @@ const promptAndSetSpace = async (Data) => {
     spaceSelect: { name, id },
   } = await prompt(question)
 
-  configstore.set('currentSpaceName', name)
-  configstore.set('currentSpaceId', id)
+  headLessConfigStore().set('currentSpaceName', name)
+  headLessConfigStore().set('currentSpaceId', id)
 
   feedback({ type: 'success', message: `${name} set` })
 }
 /**
  *
- * @param {String?} spaceName Name of space
+ * @param {String?} space_name Name of space
  * @returns
  */
-const use = async (spaceName) => {
+const use = async (space_name) => {
   // check space is linked with block
-  const currentSpaceName = configstore.get('currentSpaceName')
+  const currentSpaceName = headLessConfigStore().get('currentSpaceName')
 
   const configPath = path.resolve(BB_CONFIG_NAME)
-  const { error } = await ConfigFactory.create(configPath)
+  const { error, manager: configManager } = await ConfigFactory.create(configPath)
   if (error) {
     if (error.type !== 'OUT_OF_CONTEXT') throw error
-  } else if (currentSpaceName) {
-    console.log(chalk.dim(`current space is ${currentSpaceName}`))
-    feedback({ type: 'error', message: `Switching spaces is not allowed inside package context` })
-    process.exit(0)
+    throw new Error('Please run the command inside package context ')
+  } else if (configManager instanceof BlockConfigManager) {
+    throw new Error('Please run the command inside package context ')
   }
 
-  if (spaceName && currentSpaceName === spaceName) {
-    feedback({ type: 'info', message: `${spaceName} is already set` })
+  if (space_name && currentSpaceName === space_name) {
+    feedback({ type: 'info', message: `${space_name} is already set` })
     process.exit(0)
   }
 
@@ -72,20 +73,20 @@ const use = async (spaceName) => {
      * @type {import('../utils/jsDoc/types').spaceDetails}
      */
     const Data = res.data.data
-    if (!spaceName) {
+    if (!space_name) {
       // if space name is not given, prompt the user with available space names
       await promptAndSetSpace(Data)
       return
     }
-    const spaceDetails = Data.filter((v) => v.space_name === spaceName)[0]
+    const spaceDetails = Data.filter((v) => v.space_name === space_name)[0]
     if (!spaceDetails) {
       // if User given space name is not present, prompt the user with available space names
-      feedback({ type: 'error', message: `${spaceName} doesn't exist` })
+      feedback({ type: 'error', message: `${space_name} doesn't exist` })
       await promptAndSetSpace(Data)
     } else {
-      configstore.set('currentSpaceName', spaceDetails.space_name)
-      configstore.set('currentSpaceId', spaceDetails.space_id)
-      feedback({ type: 'success', message: `${spaceName} set` })
+      headLessConfigStore().set('currentSpaceName', spaceDetails.space_name)
+      headLessConfigStore().set('currentSpaceId', spaceDetails.space_id)
+      feedback({ type: 'success', message: `${space_name} set` })
     }
   } catch (err) {
     feedback({ type: 'error', message: err.message })